requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.6.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.0.0
idna==3.10
iniconfig==2.1.0
numpy==1.24.4
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
pluggy==1.5.0
pytest==8.3.5
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.1
